Slight design change
Fixed centering problems in css on main page for project cards and on contact page - the urge to switch to React is fierce, and will be acted upon in due time. Also, regular CSS coming from TailwindCSS is the equivalent of skipping stones on a pond with a smartphone in your pocket.